pyproject.toml
[tool.poetry]
name = "hw-6"
version = "0.1.0"
description = ""
authors = ["OleksandrChorniy <sahacherniy94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
faker = "^19.12.0"
connect = "^0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
